backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# ADB & Device Control
adbutils==2.3.0
pure-python-adb==0.3.0.dev0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# WebSocket
python-socketio==5.11.1
websockets==12.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
asyncio==3.4.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
